Move secure logic inside existing repo (#36)
* move encryption feature inside existing Repositories and remove the unnecessary Secure project

* bump version to 6.0.1

* rel notes
