+++ src/Evento.Repository/ICryptoService.cs
@@ -1,0 +1,8 @@
+﻿namespace Evento.Repository
+{
+    public interface ICryptoService
+    {
+        byte[] Encrypt(string plainText, string key);
+        string Decrypt(byte[] cipherText, string key);
+    }
+}
