+++ src/Evento.Repository/IKeyReader.cs
@@ -1,0 +1,7 @@
+﻿namespace Evento.Repository
+{
+    public interface IKeyReader
+    {
+        string Get(string id);
+    }
+}
