--- src/Evento.Secure.Repository/ICryptoService.cs
@@ -1,6 +1,0 @@
-﻿namespace Evento.Secure.Repository;
-public interface ICryptoService
-{
-    byte[] Encrypt(string plainText, string key);
-    string Decrypt(byte[] cipherText, string key);
-}
