--- src/Evento.Secure.Repository/IKeyReader.cs
@@ -1,5 +1,0 @@
-﻿namespace Evento.Secure.Repository;
-public interface IKeyReader
-{
-    string Get(string id);
-}
