+++ src/Evento.sln
@@ -15,8 +15,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Evento.Repository.Grpc", "E
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Evento.TestClient.Grpc", "Evento.TestClient.Grpc\Evento.TestClient.Grpc.csproj", "{A96B7280-279B-46ED-9AAA-6A88EA46AB80}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Evento.Secure.Repository", "Evento.Secure.Repository\Evento.Secure.Repository.csproj", "{C755D9F4-7659-4037-ACFD-CC66AC0D3399}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -47,10 +45,6 @@ Global
 		{A96B7280-279B-46ED-9AAA-6A88EA46AB80}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A96B7280-279B-46ED-9AAA-6A88EA46AB80}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A96B7280-279B-46ED-9AAA-6A88EA46AB80}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C755D9F4-7659-4037-ACFD-CC66AC0D3399}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C755D9F4-7659-4037-ACFD-CC66AC0D3399}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C755D9F4-7659-4037-ACFD-CC66AC0D3399}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C755D9F4-7659-4037-ACFD-CC66AC0D3399}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
